+++ falcosidekick-k8s-operator/terraform/main.tf
@@ -1,0 +1,19 @@
+# Copyright 2025 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+resource "juju_application" "falcosidekick-k8s" {
+  name       = var.app_name
+  model_uuid = var.model_uuid
+
+  charm {
+    name     = "falcosidekick-k8s"
+    channel  = var.channel
+    revision = var.revision
+    base     = var.base
+  }
+
+  config             = var.config
+  constraints        = var.constraints
+  units              = var.units
+  storage_directives = var.storage
+}