+++ falcosidekick-k8s-operator/terraform/outputs.tf
@@ -1,0 +1,13 @@
+# Copyright 2025 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+output "app_name" {
+  description = "Name of the deployed application."
+  value       = juju_application.falcosidekick-k8s.name
+}
+
+output "endpoints" {
+  value = {
+    ingress = "ingress"
+  }
+}