+++ falcosidekick-k8s-operator/terraform/versions.tf
@@ -1,0 +1,12 @@
+# Copyright 2025 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+terraform {
+  required_version = ">= 1.14.0"
+  required_providers {
+    juju = {
+      source  = "juju/juju"
+      version = ">= 1.1.1"
+    }
+  }
+}