+++ falcosidekick-k8s-operator/tests/unit/test_state.py
@@ -1,0 +1,8 @@
+# Copyright 2025 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+"""Unit tests for Falco state module."""
+
+
+class TestCharmState:
+    """Test CharmState class."""